requirements.txt
alembic==1.8.1
backports.zoneinfo==0.2.1
certifi
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
Mako==1.2.3
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.0
psycopg2-binary==2.9.4
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
schedule==1.1.0
scikit-learn==1.1.2
scipy==1.9.2
six
SQLAlchemy==1.4.41
threadpoolctl==3.1.0
tzdata==2022.6
tzlocal==4.2
Werkzeug==2.2.2
wincertstore==0.2
xgboost==1.6.2
zipp==3.9.0
